ref: Move interfaces
