+++ src/main/java/eu/gir/girsignals/signalbox/config/HVSignalConfig.java
@@ -9,7 +9,6 @@ import eu.gir.girsignals.EnumSignals.HP_TYPE;
 import eu.gir.girsignals.EnumSignals.VR;
 import eu.gir.girsignals.EnumSignals.ZS32;
 import eu.gir.girsignals.SEProperty;
-import eu.gir.girsignals.blocks.ISignalAutoconfig;
 import eu.gir.girsignals.blocks.signals.SignalHV;
 import eu.gir.girsignals.tileentitys.SignalTileEnity;
